Fix code formatting/style issues ...
... including, but not limited to:
- Useless casts
- Unused functions & macros
- K&R-style function declarations
- private functions not marked as static
- printf() format truncation
- signed/unsigned comparisons
- redundant declarations
- unprototyped global functions
- variable shadowing

(detected by various non-default GCC 8 warning options)

+++ glxdemos/offset.c
@@ -348,7 +348,7 @@ query_extension(char* extName) {
     char *p = (char *) glGetString(GL_EXTENSIONS);
     char *end = p + strlen(p);
     while (p < end) {
-        int n = strcspn(p, " ");
+        size_t n = strcspn(p, " ");
         if ((strlen(extName) == n) && (strncmp(extName, p, n) == 0))
             return GL_TRUE;
         p += (n + 1);
